requirements.txt
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
Django==1.10.6
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-haystack==2.6.1
djangorestframework==3.6.4
httpie==0.9.9
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Markdown==2.6.9
olefile==0.44
parso==0.1.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
Pygments==2.2.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
